pentagi_client_pkg/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pentagi-client"
version = "0.1.0"
description = "Python client for the PentAGI AI penetration testing platform"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31",
    "websocket-client>=1.7",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[project.scripts]
pentagi = "pentagi_client.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-mock>=3", "responses>=0.25"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pentagi_client*"]
